renames videojs5-hlsjs-source-handler.js -> videojs-hlsjs-source-handler.js
