+++ lib/main.js
@@ -1,4 +1,4 @@
 var Hls = require('hls.js');
-var attachVideojsStreamrootProvider = require('./videojs5-hlsjs-source-handler.js');
+var attachVideojsStreamrootProvider = require('./videojs-hlsjs-source-handler.js');
 
 attachVideojsStreamrootProvider(window, window.videojs, Hls);
