+++ lib/videojs-hlsjs-source-handler.js
